Feature/foss 459 fixes during firmware debug

+++ src/idm_nvme_io.h
@@ -135,7 +135,7 @@ int nvme_idm_write(nvmeIdmRequest *request_idm);
 int nvme_idm_write_init(char *lock_id, int mode, char *host_id, char *drive,
                         uint64_t timeout, nvmeIdmRequest *request_idm);
 
-int _nvme_idm_cmd_check_status(int status, int opcode_idm);
+int _nvme_idm_cmd_check_status(int status, uint8_t opcode_idm);
 int _nvme_idm_cmd_init(nvmeIdmRequest *request_idm, uint8_t opcode_nvme);
 int _nvme_idm_cmd_send(nvmeIdmRequest *request_idm);
 int _nvme_idm_data_init_wrt(nvmeIdmRequest *request_idm);
